pyproject.toml
[tool.poetry]
name = "tft-analysis"
version = "0.1.0"
description = ""
authors = ["Max Gonzih <gonzih@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.4"
pydantic = "^1.10.1"
PyYAML = "^6.0"
coloredlogs = "^15.0.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
